+++ CHANGELOG.md
@@ -1,2 +1,2 @@
-* Removes disabled items from loot tables.
-* Fixed crash when editing config without Cloth Config.
+* Add default config rule.
+* Per-dimension interaction blacklist.